chore: merge some modules and export Context type
