+++ packages/core/src/index.ts
@@ -19,6 +19,7 @@ export type {
   RsbuildPluginAPI,
 } from './types';
 export type {
+  Context as RsbuildContext,
   RsbuildMode,
   RsbuildEntry,
   RsbuildTarget,
