+++ packages/core/src/rspack-provider/config.ts
@@ -7,8 +7,9 @@ import {
   getDefaultSourceConfig,
   getDefaultSecurityConfig,
   getDefaultPerformanceConfig,
+  type RsbuildConfig,
+  type NormalizedConfig,
 } from '@rsbuild/shared';
-import type { RsbuildConfig } from '../../types';
 
 export const createDefaultConfig = (): RsbuildConfig => ({
   dev: getDefaultDevConfig(),
@@ -22,3 +23,14 @@ export const createDefaultConfig = (): RsbuildConfig => ({
 
 export const withDefaultConfig = (config: RsbuildConfig) =>
   mergeRsbuildConfig<RsbuildConfig>(createDefaultConfig(), config);
+
+/** #__PURE__
+ * 1. May used by multiple plugins.
+ * 2. Object value that should not be empty.
+ * 3. Meaningful and can be filled by constant value.
+ */
+export const normalizeConfig = (config: RsbuildConfig): NormalizedConfig =>
+  mergeRsbuildConfig<NormalizedConfig>(
+    createDefaultConfig() as NormalizedConfig,
+    config as NormalizedConfig,
+  );
