+++ packages/core/src/rspack-provider/css-modules-typescript-pre-loader/postcss-icss-extract-plugin.ts
