+++ packages/core/src/rspack-provider/index.ts
@@ -1,5 +1,4 @@
-export { getRspackVersion } from './shared/rspackVersion';
+export { getRspackVersion } from './shared';
 export { rspackProvider } from './provider';
 export type { RspackProvider } from './provider';
-
 export type { Rspack, RspackConfig } from '@rsbuild/shared';
