--- packages/core/src/rspack-provider/shared/constants.ts
@@ -1,1 +1,0 @@
-export const BUILTIN_LOADER = 'builtin:';
