--- packages/core/src/rspack-provider/shared/fs.ts
@@ -1,15 +1,0 @@
-import { join } from 'path';
-import {
-  getSharedPkgCompiledPath,
-  SharedCompiledPkgNames,
-} from '@rsbuild/shared';
-import { fse } from '@rsbuild/shared';
-
-export const getCompiledPath = (packageName: string) => {
-  const providerCompilerPath = join(__dirname, '../../compiled', packageName);
-  if (fse.existsSync(providerCompilerPath)) {
-    return providerCompilerPath;
-  } else {
-    return getSharedPkgCompiledPath(packageName as SharedCompiledPkgNames);
-  }
-};
