--- packages/core/src/rspack-provider/shared/index.ts
@@ -1,3 +1,0 @@
-export * from './fs';
-export * from './constants';
-export * from './rspackVersion';
