+++ packages/document/docs/en/api/javascript-api/types.mdx
@@ -12,14 +12,14 @@ import type { RsbuildInstance } from '@rsbuild/core';
 let rsbuild: RsbuildInstance;
 ```
 
-## Context
+## RsbuildContext
 
 The type of the context property in the Rsbuild instance.
 
 ```ts
-import type { Context } from '@rsbuild/core';
+import type { RsbuildContext } from '@rsbuild/core';
 
-const context: Context = rsbuild.context;
+const context: RsbuildContext = rsbuild.context;
 ```
 
 ## RsbuildPlugin
