+++ packages/document/docs/zh/api/javascript-api/types.mdx
@@ -12,14 +12,14 @@ import type { RsbuildInstance } from '@rsbuild/core';
 let rsbuild: RsbuildInstance;
 ```
 
-## Context
+## RsbuildContext
 
 Rsbuild 实例中 context 属性的类型定义。
 
 ```ts
-import type { Context } from '@rsbuild/core';
+import type { RsbuildContext } from '@rsbuild/core';
 
-const context: Context = rsbuild.context;
+const context: RsbuildContext = rsbuild.context;
 ```
 
 ## RsbuildPlugin
